moving static out of handlers
