+++ src/static.js
